added more tasks before ui
