+++ test/git-tasks.js
@@ -77,12 +77,38 @@ test.serial('should fail when remote history differs', async t => {
 	t.true(SilentRenderer.tasks.some(task => task.title === 'Check remote history' && task.hasFailed()));
 });
 
-test.serial('checks should pass when publishing from master, working tree is clean and remote history not different', async t => {
+test.serial('should fail when release branch is not specified, current branch is not main/master and publishing from any branch not permitted', async t => {
 	execaStub.createStub([
 		{
 			command: 'git symbolic-ref --short HEAD',
 			exitCode: 0,
-			stdout: 'master'
+			stdout: 'feature'
+		}
+	]);
+	await t.throwsAsync(run(testedModule({})),
+		{message: 'Not on `main`/`master` branch. Use --any-branch to publish anyway, or set a different release branch using --branch.'});
+	t.true(SilentRenderer.tasks.some(task => task.title === 'Check current branch' && task.hasFailed()));
+});
+
+test.serial('should fail when current branch is not the specified release branch and publishing from any branch not permitted', async t => {
+	execaStub.createStub([
+		{
+			command: 'git symbolic-ref --short HEAD',
+			exitCode: 0,
+			stdout: 'feature'
+		}
+	]);
+	await t.throwsAsync(run(testedModule({branch: 'release'})),
+		{message: 'Not on `release` branch. Use --any-branch to publish anyway, or set a different release branch using --branch.'});
+	t.true(SilentRenderer.tasks.some(task => task.title === 'Check current branch' && task.hasFailed()));
+});
+
+test.serial('should not fail when current branch not master and publishing from any branch permitted', async t => {
+	execaStub.createStub([
+		{
+			command: 'git symbolic-ref --short HEAD',
+			exitCode: 0,
+			stdout: 'feature'
 		},
 		{
 			command: 'git status --porcelain',
@@ -95,5 +121,21 @@ test.serial('checks should pass when publishing from master, working tree is cle
 			stdout: ''
 		}
 	]);
-	await t.notThrowsAsync(run(testedModule({})));
+	await run(testedModule({anyBranch: true}));
+	t.false(SilentRenderer.tasks.some(task => task.title === 'Check current branch'));
 });
+
+test.serial('should fail when git version does not match range in `package.json`', async t => {
+	execaStub.createStub([
+		{
+			command: 'git version',
+			exitCode: 0,
+			stdout: 'git version 1.0.0'
+		}
+	]);
+	const depRange = require('../package.json').engines.git;
+	await t.throwsAsync(run(testedModule()),
+		{message: `Please upgrade to git${depRange}`});
+	t.true(SilentRenderer.tasks.some(task => task.title === 'Check git version' && task.hasFailed()));
+});
+
