+++ test/prerequisite-tasks.js
@@ -155,72 +155,6 @@ test.serial('private package: should skip task `verify user is authenticated`', 
 	t.true(SilentRenderer.tasks.some(task => task.title === 'Verify user is authenticated' && task.isSkipped()));
 });
 
-test.serial('should fail when git version does not match range in `package.json`', async t => {
-	execaStub.createStub([
-		{
-			command: 'git version',
-			exitCode: 0,
-			stdout: 'git version 1.0.0'
-		}
-	]);
-	const depRange = require('../package.json').engines.git;
-	await t.throwsAsync(run(testedModule('2.0.0', {name: 'test', version: '1.0.0'}, {yarn: false})),
-		{message: `Please upgrade to git${depRange}`});
-	t.true(SilentRenderer.tasks.some(task => task.title === 'Check git version' && task.hasFailed()));
-});
-
-test.serial('should fail when release branch is not specified, current branch is not main/master and publishing from any branch not permitted', async t => {
-	execaStub.createStub([
-		{
-			command: 'git symbolic-ref --short HEAD',
-			exitCode: 0,
-			stdout: 'feature'
-		}
-	]);
-	await t.throwsAsync(run(testedModule('2.0.0', {name: 'test', version: '1.0.0', private: true}, {yarn: false})),
-		{message: 'Not on `main`/`master` branch. Use --any-branch to publish anyway, or set a different release branch using --branch.'});
-	t.true(SilentRenderer.tasks.some(task => task.title === 'Check current branch' && task.hasFailed()));
-});
-
-test.serial('should fail when current branch is not the specified release branch and publishing from any branch not permitted', async t => {
-	execaStub.createStub([
-		{
-			command: 'git symbolic-ref --short HEAD',
-			exitCode: 0,
-			stdout: 'feature'
-		}
-	]);
-	await t.throwsAsync(run(testedModule('2.0.0', {name: 'test', version: '1.0.0', private: true}, {yarn: false, branch: 'release'})),
-		{message: 'Not on `release` branch. Use --any-branch to publish anyway, or set a different release branch using --branch.'});
-	t.true(SilentRenderer.tasks.some(task => task.title === 'Check current branch' && task.hasFailed()));
-});
-
-test.serial('should not fail when current branch not master and publishing from any branch permitted', async t => {
-	execaStub.createStub([
-		{
-			command: 'git symbolic-ref --short HEAD',
-			exitCode: 0,
-			stdout: 'feature'
-		},
-		{
-			command: 'git status --porcelain',
-			exitCode: 0,
-			stdout: ''
-		},
-		{
-			command: 'git rev-list --count --left-only @{u}...HEAD',
-			exitCode: 0,
-			stdout: ''
-		},
-		{
-			command: 'git rev-parse --quiet --verify refs/tags/v2.0.0',
-			stdout: ''
-		}
-	]);
-	await run(testedModule('2.0.0', {name: 'test', version: '1.0.0', private: true}, {yarn: false, anyBranch: true}));
-	t.true(SilentRenderer.tasks.some(task => task.title === 'Check current branch' && task.isSkipped()));
-});
-
 test.serial('should fail when version is invalid', async t => {
 	await t.throwsAsync(run(testedModule('DDD', {name: 'test', version: '1.0.0'}, {yarn: false})),
 		{message: `Version should be either ${version.SEMVER_INCREMENTS.join(', ')}, or a valid semver version.`});
